Better types for Sanity Articles
